Allow for unicode characters in messages
A unicode error was being thrown because the string that was being
formatted-into was not unicode. It's now unicode, which appears to fix
the issue of unicode messages causing errors.

If you want to throw an emoji in your SNS message and process it with
this library, go for it!