Unicode messages, README cleanup, python 3, expanded travis, & more! (#1)
* Updating README to fix spacing issues

Maintaining consistency for spacing across the entire README. Changing
the instructions to use `bash` because that's not actually run in
python, and removing an unnecessary `=` from the header.

* Removing developer-specific Visual Studio folder

The .vscode file is hard coded with paths specific to (presumably) the
author of the package, and isn't necessary for the package.

* Fleshing out .gitignore to cover more possible files

As well as adding the .vscode folder from before. So we don't run into
the same problem as before.

* Have Travis run more tests with more supported versions

* Update views.py to be support python3

* Add cffi as a requirement, make manage.py executable

* Fix some typos on travis, bump setuptools, no 3.3

Python 3.3 just doesn't work. It isn't supported in anything but 1.8
anyways and really isn't worth the effort here.

* Bump version to 0.1.2 to make releasing to pypi easy

* Add `self` to the process_message example

* Allow for unicode characters in messages

A unicode error was being thrown because the string that was being
formatted-into was not unicode. It's now unicode, which appears to fix
the issue of unicode messages causing errors.

If you want to throw an emoji in your SNS message and process it with
this library, go for it!